Split Data Module - remote and repository
