+++ data/remote/src/main/java/com/ameri/remote/Constants.kt
@@ -1,4 +1,4 @@
-package com.ameri.data
+package com.ameri.remote
 
 
 //Server
