+++ data/remote/src/main/java/com/ameri/remote/network/ApiServices.kt
@@ -1,8 +1,8 @@
-package com.ameri.data.network
+package com.ameri.remote.network
 
 
-import com.ameri.data.*
 import com.ameri.domain.models.ResponseData
+import com.ameri.remote.*
 import retrofit2.Response
 import retrofit2.http.GET
 import retrofit2.http.Query
