+++ data/repository/src/main/java/com/ameri/repository/NewsRepositoryImpl.kt
@@ -1,8 +1,8 @@
-package com.ameri.data.repository
+package com.ameri.repository
 
-import com.ameri.data.network.ApiServices
 import com.ameri.domain.models.ResponseData
 import com.ameri.domain.repository.NewsRepository
+import com.ameri.remote.network.ApiServices
 import retrofit2.Response
 import javax.inject.Inject
