+++ settings.gradle.kts
@@ -24,3 +24,5 @@ include(":app")
 include(":presentation")
 include(":domain")
 include(":data")
+include(":data:remote")
+include(":data:repository")
